pyproject.toml
[tool.poetry]
name = "pet-hotel"
version = "0.1.0"
description = ""
authors = ["nisrinafri <febriyantinisrina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["fastapi"], version = "^0.234.0"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pydeps = "^1.12.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
